컴파일 에러 fix: GUI.depth는 int (float→int 변환 누락)
DrawClueTooltip에서 prevDepth 타입을 float → int로.
GUI.depth는 int 타입이라 float에 저장 후 복원 시 implicit cast 실패.

+++ Assets/_Project/Scripts/Systems/FacilityHUD.cs
@@ -1960,7 +1960,7 @@ namespace ForTheCompany.Systems
             var tipR = new Rect(tx, ty, tipW, tipH);
 
             // GUI.depth 음수로 위에 그리도록
-            float prevDepth = GUI.depth;
+            int prevDepth = GUI.depth;
             GUI.depth = -500;
 
             UITheme.DrawRect(tipR, UITheme.Bg1);
